Changed SWIG dependendence to optional, and set up things so that the autogenerated XS code (*_wrap.c) is included in the distribution, but is not under source control.
Thanks to Eric Wilhelm for this removal of a barrier to entry. Now it is set up so that when a developer clones a git
repository, the autogenerated files will not be present and therefore the Build.PL script will attempt to generate them
from the SWIG interface files. But for people installing via CPAN, the *_wrap.c files will be present and therefore
end users will not be forced to have SWIG to install Math::GSL.