In instantiation of GSLBuilder in Build.PL, the value for the parameter include_dirs were corrected to an anonymous empty array, instead of an empty string.
There were a build failure where gcc gets confused and 'loses' one
of the supplied -I flags (the one specifying the directory
containing the core perl header files), resulting in gcc not
finding a number of very fundamental header files for perl modules.

This is caused by a single '-I' flag appearing before the 'lost' -I
flag.  That lone '-I' flag is the result of trying to use the -I
flag to add a directory consisting of the empty string to the list
of directories to search for include files.

The fix is to prevent that bogus empty string directory to enter
the build logic in the first place:

In the instantiation of the GSLBuilder class, I changed the
value passed for the 'include_dirs' from an empty string to a
(reference to) an empty array.  This prevents the lone '-I' flag
to be passed to gcc, making gcc properly understand the rest
of the command line.

(The GSLBuilder code can be amended to actually fail when
'include_dirs' is given an empty string, and in that failure,
tell that 'include_dirs' should be given an empty list if
one wants to state that there should be no include directories)