Still fighting the loss of precision dragons
RT66882 is an interesting bug. A single special function test fails due to
loss of precision, only when DUSELONGDOUBLE is defined in Perl. It seems
to have the same behavior across a wide number of Perl versions, from 5.6
to 5.13.x. Why are long doubles causing this loss of precision? Haven't
the slightest, currently. It could be a bug in Perl 5, GSL or Math::GSL.

It's a bug hunt.