Attempt to fix non-deterministic errors when building GSL in Travis
GSL installation script has been reworked to:

* isolate errors and provide meaningful error messages
* ensure that GSL downloads and builds are only done as necessary
* switch from ftp.gnu.org to mirrors.kernel.org.

Tests indicated that fetching the source from the GNU ftp site was the
source of failures.  Switching to the mirrors.kernel.org site seems
to have resolved this issue.  The reworked code is just icing on the cake.

There's probably too much logic in here, and the result of Travis'
caching seems different to me than what is described by the documentation.