Fix Version.i
Version.i included gsl_version.h which had defined GSL_VERSION as 2.6+.
This caused the test t/00-load.t to fail with error "Invalid version
format (non-numeric data)". This was because of the "+" at the end of
"2.6+". To fix this, we can instead defined the version as "2.6" in
system.i and avoid include "gsl_version.h" in Version.i.