Add gsl_blas_types.h to Matrix.i
In GSL 2.6 gsl_matrix_tricpy() has got a new signature that depends on
include file gsl_blas_types.h. This include file is not included from
gsl_matrix_double.h in GSL 2.5. In order for the generated C files for
GSL 2.5 to compile we need to include gsl_blas_types.h in the Matrix.i
and MatrixComplex.i files.