Add version information from swig command line.
Adds macros definitions for the GSL version that is currently being
built from the command line in compile_swig() in GSLBuilder.pm. This
will enable the interface files to check which GSL version they are
supposed to be used for. This feature can be used to include new
include files into the inteface file based on the GSL version or
it can be used for example to exclude or rewrite definition. This
is an alternative/extension of the rename.i framework in Ver2Func.pm.