Fix rewrite of gsl_spmatrix.h in GSL 2.6
In GSL 2.6 gsl_spmatrix.h was refactored and a lot of new include files
was introduced. This is a quick fix to have this compile with GSL
versions less than 2.6.

I think, ideally all the headers for GSL 2.5 should have been included
in "./include/legacy/gsl-2.5" and not only gsl_spmatrix.h. But this
at least seems to work, and we can add more files later if needed.