Fix Version.i to use correct version.
We need to pass the current build version to Version.i. It needs to set
the correct version for GSL_VERSION. This is not the value as given in
system.i since that is the version that Math::GSL distribution was
built with and not the version on the system where the user is
installing the distribution.